+++ src/utils/index.js
@@ -7,6 +7,7 @@ const generateToken = require('./generateToken');
 const { makeAbsolute, actionMatches, parseActionPayload } = require('./pathUtils');
 const { makeQuickReplies, quickReplyAction } = require('./quickReplies');
 const { replaceDiacritics, tokenize } = require('./tokenizer');
+const compileWithState = require('./compileWithState');
 
 module.exports = {
     replaceDiacritics,
@@ -16,5 +17,6 @@ module.exports = {
     makeAbsolute,
     actionMatches,
     parseActionPayload,
-    generateToken
+    generateToken,
+    compileWithState
 };