+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.13.0] 2021-03-26
+
+### Added
+
+- `Responder.trackEvent()` method for tracking purposes
+- `ReturnSender.tracking` object with tracked events
+- `tracking` property in `Processor#interaction` event
+- `ai.wingbot.trackingEvent` plugin
+
 ## [3.12.0] 2021-03-26
 
 ### Added